Added fetchAllCarts and cartForUser endpoints

+++ src/main/java/com/ecommerce/project/model/CartItem.java
@@ -5,9 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.ArrayList;
-import java.util.List;
-
 @Entity
 @Data
 @Table(name="cart_items")
@@ -26,8 +23,7 @@ public class CartItem {
     @ManyToOne
     @JoinColumn(name = "product_id")
     private  Product product;
-
-    private Integer productPrice;
+    private Double productPrice;
     private Integer quantity;
     private Double discount;
 }
